Locked Doors Vol 1
Updated the map to make the bridges impassable, Created a bridge Level. Updated the Door Script to be lockable, save it's lock state in level data so that it remains in level re-entry, and unlocks when the player interacts with it while the specified key is in the player's possession. Created a Level Loader script so that the level borders can transport dracula to scenes besides the map. Finally added functionality for the right vs left spawning mechanic.